+++ modules/cpdos/path_traversal.py
@@ -15,7 +15,7 @@ except:
 
 logger = configure_logger(__name__)
 
-
+VULN_NAME = "Path Traversal"
 CONTENT_DELTA_RANGE = 500
 
 
@@ -31,10 +31,10 @@ def verify(req_main, url, url_cb, url_test, completed_path, range_exlusion, p, s
 
         req_cb = s.get(url_cb, verify=False, timeout=10, allow_redirects=False)
         #print(f"req_cb.status_code: {req_cb.status_code} | req_verify.status_code: {req_verify.status_code} | req_main.status_code: {req_main.status_code}")
-        if req_cb.status_code == req_verify.status_code and req_cb.status_code != req_main.status_code:
-            print(f" {Identify.confirmed} | CPDoSError {req_main.status_code} > {req_cb.status_code} | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
-        elif len(req_cb.content) not in range_exlusion and req_test.status_code not in [403, 401, 429]:
-            print(f" {Identify.confirmed} | CPDoSError {len(req_main.content)}b > {len(req_cb.content)}b | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
+        if req_cb.status_code == req_verify.status_code and req_cb.status_code != req_main.status_code and req_cb.status_code not in [403, 401, 429]:
+            print(f" {Identify.confirmed} | {VULN_NAME} {req_main.status_code} > {req_cb.status_code} | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
+        elif len(req_cb.content) not in range_exlusion and req_cb.status_code not in [403, 401, 429]:
+            print(f" {Identify.confirmed} | {VULN_NAME} {len(req_main.content)}b > {len(req_cb.content)}b | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
     except requests.Timeout:
         #print(f"request timeout {url} {p}")
         pass
@@ -48,6 +48,7 @@ def path_traversal_check(url, s, req_main, authent):
     try:
         range_exlusion = range(len(req_main.content) - CONTENT_DELTA_RANGE, len(req_main.content) + CONTENT_DELTA_RANGE)
         paths = [
+        "\\",
         "cc\\..\\",
         "cc/../",
         "cc/%2e%2e%2f"
@@ -65,10 +66,10 @@ def path_traversal_check(url, s, req_main, authent):
 
             req_test = s.get(url_test, verify=False, timeout=10, allow_redirects=False)
             if req_test.status_code != req_main.status_code and req_test.status_code not in [403, 401, 429]:
-                print(f" {Identify.behavior} | CPDoSError {req_main.status_code} > {req_test.status_code} | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
+                print(f" {Identify.behavior} | {VULN_NAME} {req_main.status_code} > {req_test.status_code} | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
                 verify(req_main, url, url_cb, url_test, completed_path, range_exlusion, p, s)
             elif len(req_test.content) not in range_exlusion and req_test.status_code not in [403, 401, 429]:
-                print(f" {Identify.behavior} | CPDoSError {len(req_main.content)}b > {len(req_test.content)}b | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
+                print(f" {Identify.behavior} | {VULN_NAME} {len(req_main.content)}b > {len(req_test.content)}b | \033[34m{url_cb}\033[0m | PAYLOAD: {url_test}")
                 verify(req_main, url, url_cb, url_test, completed_path, range_exlusion, p, s)
     except requests.Timeout:
         #print(f"request timeout {url} {p}")